backend/skills_library/info_collection/delegate_info_recognition/requirements.txt
python-dotenv>=1.0.0
pyyaml>=6.0
pdf2image>=1.16.0
Pillow>=10.0.0
dashscope>=1.14.0
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
httpx>=0.25.0
